pyproject.toml
[tool.poetry]
name = "api1"
version = "0.1.0"
description = ""
authors = ["Nate Allen <nallen@directionsresearch.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.95.0"}
psycopg2 = "^2.9.5"
fastapi-sqlalchemy = "^0.2.1"
pydantic = "^1.10.7"
alembic = "^1.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
